feat: restructure tests according to src structure
