+++ tests/evaluation/evaluation/test_argilla_evaluator.py
@@ -27,13 +27,63 @@ from intelligence_layer.evaluation import (
     Runner,
     SuccessfulExampleOutput,
 )
-from tests.conftest import (
+from tests.evaluation.conftest import (
     DummyStringEvaluation,
     DummyStringInput,
     DummyStringOutput,
     DummyStringTask,
 )
-from tests.evaluation.conftest import StubArgillaClient
+
+
+class StubArgillaClient(ArgillaClient):
+    _expected_workspace_id: str
+    _expected_fields: Sequence[Field]
+    _expected_questions: Sequence[Question]
+    _datasets: dict[str, list[RecordData]] = {}
+    _score = 3.0
+
+    def ensure_dataset_exists(
+        self,
+        workspace_id: str,
+        dataset_name: str,
+        fields: Sequence[Field],
+        questions: Sequence[Question],
+    ) -> str:
+        if workspace_id != self._expected_workspace_id:
+            raise Exception("Incorrect workspace id")
+        elif fields != self._expected_fields:
+            raise Exception("Incorrect fields")
+        elif questions != self._expected_questions:
+            raise Exception("Incorrect questions")
+        dataset_id = str(uuid4())
+        self._datasets[dataset_id] = []
+        return dataset_id
+
+    def add_record(self, dataset_id: str, record: RecordData) -> None:
+        if dataset_id not in self._datasets:
+            raise Exception("Add record: dataset not found")
+        self._datasets[dataset_id].append(record)
+
+    def evaluations(self, dataset_id: str) -> Iterable[ArgillaEvaluation]:
+        dataset = self._datasets.get(dataset_id)
+        assert dataset
+        return [
+            ArgillaEvaluation(
+                example_id=record.example_id,
+                record_id="ignored",
+                responses={"human-score": self._score},
+                metadata=dict(),
+            )
+            for record in dataset
+        ]
+
+    def split_dataset(self, dataset_id: str, n_splits: int) -> None:
+        raise NotImplementedError
+
+
+@fixture
+def stub_argilla_client() -> StubArgillaClient:
+    return StubArgillaClient()
 
 
 class DummyStringTaskArgillaEvaluationLogic(
