+++ tests/evaluation/run/test_run.py
@@ -75,11 +75,11 @@ def test_run_evaluation(
         [
             "",
             "--eval-logic",
-            "tests.evaluation.test_run.DummyEvaluationLogic",
+            "tests.evaluation.run.test_run.DummyEvaluationLogic",
             "--aggregation-logic",
-            "tests.evaluation.test_run.DummyAggregationLogic",
+            "tests.evaluation.run.test_run.DummyAggregationLogic",
             "--task",
-            "tests.evaluation.test_run.DummyTask",
+            "tests.evaluation.run.test_run.DummyTask",
             "--dataset-repository-path",
             str(dataset_path),
             "--dataset-id",
@@ -112,11 +112,11 @@ def test_run_evaluation_with_task_with_client(
         [
             "",
             "--eval-logic",
-            "tests.evaluation.test_run.DummyEvaluationLogic",
+            "tests.evaluation.run.test_run.DummyEvaluationLogic",
             "--aggregation-logic",
-            "tests.evaluation.test_run.DummyAggregationLogic",
+            "tests.evaluation.run.test_run.DummyAggregationLogic",
             "--task",
-            "tests.evaluation.test_run.DummyTaskWithClient",
+            "tests.evaluation.run.test_run.DummyTaskWithClient",
             "--dataset-repository-path",
             str(dataset_path),
             "--dataset-id",
