+++ tests/examples/classify/test_classify.py
