+++ tests/examples/classify/test_embedding_based_classify.py
