+++ tests/examples/classify/test_prompt_based_classify.py
