+++ tests/examples/classify/test_prompt_based_classify_with_definitions.py
