+++ tests/examples/search/test_expand_chunk.py
