+++ tests/examples/search/test_search.py
